requirements.txt
# Journal Analyzer — API and Shiny app
fastapi
uvicorn
pandas
numpy
shiny
shiny[theme]
requests
python-dotenv
supabase
sentence-transformers
openai
markdown
plotly
